Refactor : Made Components more modular
