+++ src/components/VideoCard/VideoCard.jsx
@@ -7,19 +7,14 @@ import {
 } from "../../context/Video/Watchlater";
 import styles from "./VideoCard.module.css";
 
-export default function VideoCard({ video, mini }) {
+export default function VideoCard({ video, mini, children }) {
   const { thumbnail, preview, title, creator, views } = video;
   const [options, setOptions] = useState(false);
-  const { userState, userDispatch } = useUserDetails();
-  const { watchlater } = userState;
 
   const optionsHandler = () => {
     setOptions((prev) => !prev);
   };
 
-  const watchlaterHandler = (video) => {
-    return watchlater?.some((item) => video._id === item._id);
-  };
   return (
     <div
       className={`${styles.video_card} ${
@@ -41,26 +36,7 @@ export default function VideoCard({ video, mini }) {
           options ? styles.showDialog : styles.hideDialog
         }`}
       >
-        <div
-          onClick={() =>
-            watchlaterHandler(video)
-              ? removeFromWatchlater(video, userDispatch)
-              : addToWatchlater(video, userDispatch)
-          }
-          className={`${styles.dialog} flex-row-nowrap`}
-        >
-          <span className="material-icons">watch_later</span>
-          <p>
-            {watchlaterHandler(video)
-              ? "Del from Watch Later"
-              : "Add to Watch Later"}
-          </p>
-        </div>
-        <div className={`${styles.dialog} flex-row-nowrap`}>
-          <span className="material-icons">playlist_add</span>
-          <label htmlFor="watchlater">Add to Playlist</label>
-          <input id="watchlater" type="checkbox" />
-        </div>
+        {children}
       </div>
       <Link to={`/video/${video.videoID}`}>
         <div className={styles.videoCover}>
