+++ src/components/VideoCard/VideoCard.module.css
@@ -72,25 +72,6 @@
   filter: brightness(0.9);
 }
 
-.dialog {
-  gap: 10px;
-  align-items: center;
-  justify-content: flex-start;
-  padding: 0.25 rem 0px;
-  margin-bottom: 1rem;
-  border-radius: 8px;
-
-  font-size: small;
-  font-weight: bold;
-  background-color: inherit;
-  transition: background-color var(--ANIM) 300ms;
-}
-
-.dialog:hover {
-  background-color: var(--HOVER_BG_ACTIVE);
-  transition: background-color var(--ANIM) 300ms;
-}
-
 .dialog_container {
   position: absolute;
   width: 100%;
